+++ packages/insomnia/src/account/session.ts
@@ -5,7 +5,7 @@ import { models, services } from 'insomnia-data';
 
 import { AI_PLUGIN_NAME, LLM_BACKENDS } from '../common/constants';
 import { database } from '../common/database';
-import { decryptAES } from '../runtimes/crypto/crypt-adapter';
+import { decryptAES } from '../runtimes/crypto/crypto-adapter';
 
 export interface SessionData {
   accountId: string;
