+++ packages/insomnia/src/main/ipc/main.ts
@@ -54,7 +54,7 @@ import type {
 import * as crypt from '../../account/crypt';
 import type { HiddenBrowserWindowBridgeAPI } from '../../entry.hidden-window';
 import type { PluginsBridgeAPI } from '../../plugins/bridge-types';
-import { decryptSecretValue, encryptSecretValue } from '../../runtimes/crypto/crypt-adapter';
+import { decryptSecretValue, encryptSecretValue } from '../../runtimes/crypto/crypto-adapter';
 import type { RenderedRequest } from '../../templating/types';
 import { keyPair as sealedboxKeyPair, open as sealedboxOpen } from '../../utils/sealedbox';
 import type { AnalyticsEvent } from '../analytics';
