--- packages/insomnia/src/runtimes/crypto/crypt-adapter.ts
@@ -1,3 +1,0 @@
-// Imports the renderer implementation by default.
-// esbuild node builds alias this to crypt-adapter.node via the renderer-to-node plugin.
-export { encryptSecretValue, decryptSecretValue, decryptAES } from './crypt-adapter.renderer';
