+++ packages/insomnia/src/runtimes/crypto/crypto-adapter.ts
@@ -1,0 +1,3 @@
+// Imports the renderer implementation by default.
+// esbuild node builds alias this to crypto-adapter.node via the renderer-to-node plugin.
+export { encryptSecretValue, decryptSecretValue, decryptAES } from './crypto-adapter.renderer';
