--- packages/insomnia/src/runtimes/templating/render-adapter.ts
@@ -1,3 +1,0 @@
-// Imports the renderer implementation by default.
-// esbuild node builds alias this to render-adapter.node via the renderer-to-node plugin.
-export { renderTemplate } from './render-adapter.renderer';
