+++ packages/insomnia/src/runtimes/templating/templating-adapter.node.ts
