+++ packages/insomnia/src/runtimes/templating/templating-adapter.ts
@@ -1,0 +1,3 @@
+// Imports the renderer implementation by default.
+// esbuild node builds alias this to templating-adapter.node via the renderer-to-node plugin.
+export { renderTemplate } from './templating-adapter.renderer';
